feat: move the "Queries" struct from common to backend
